Convert WinMTR GUI app to CLI/service tool (v0.0.1)
Remove the MFC GUI and turn WinMTR into a headless Win32 console
application (winmtrcli.exe) suitable for 24/7 operation as an nssm
service.

- Decouple WinMTRNet from the dialog via a new MTRConfig struct and
  replace AfxMessageBox with stderr output.
- Make WinMTRGlobal.h MFC-free (winsock2/iphlpapi/icmpapi).
- New console entry point with argv parsing and SetConsoleCtrlHandler
  for clean shutdown on Ctrl+C / service stop.
- New reporter module: continuous live .log (LogExpert-friendly) and CSV
  logging plus final text/HTML summaries; flush + _commit so the file
  size updates live.
- Auto-generate default output files (<host>_<timestamp>.log/_results.*)
  in the working directory when no output option is given.
- Retarget project to Console subsystem, platform toolset v145, link
  ws2_32/iphlpapi; drop all GUI/MFC source files.
- Update README with CLI usage and nssm service instructions.

+++ WinMTRGlobal.cpp
@@ -19,7 +19,7 @@ int gettimeofday(struct timeval* tv, struct timezone *tz)
    
    _ftime(&timebuffer);
 
-   tv->tv_sec = timebuffer.time;
+   tv->tv_sec = (long)timebuffer.time;
    tv->tv_usec = timebuffer.millitm * 1000 + 500;
    return 0;
 }
